core: move generic_core_bootcfg.c into core_mmu.c
Moves core/arch/arm/kernel/generic_core_bootcfg.c into
core/arch/arm/mm/core_mmu.c as generic_core_bootcfg.c isn't optional
anymore.
