+++ core/arch/arm/kernel/sub.mk
@@ -33,7 +33,6 @@ srcs-y += wait_queue.c
 srcs-$(CFG_PM_STUBS) += pm_stubs.c
 
 srcs-$(CFG_GENERIC_BOOT) += generic_boot.c
-srcs-y += generic_core_bootcfg.c
 ifeq ($(CFG_GENERIC_BOOT),y)
 srcs-$(CFG_ARM32_core) += generic_entry_a32.S
 srcs-$(CFG_ARM64_core) += generic_entry_a64.S
